# [SPARK-18186] Migrate HiveUDAFFunction to TypedImperativeAggregate for partial aggregation support