@liancheng 2016-10-31
Will add more details in the PR description soon.