@liancheng 2016-11-01
I can't reproduce those test failures when executing failed test cases individually. Seems that it's related to execution order. Still investigating.