@tejasapatil 2016-11-01
This will surely improve performance for the UDAFs where the data shrinks (eg. `max` as you pointed out). I am not sure if it would be better for UDAFs like `GenericUDAFCollectSet`, `GenericUDAFCollectList` where the aggregation does not shrink the data (it might be worse because of conversion cost ?).

It would be good to see some perf numbers to compare (native Spark UDFs) vs (before this change) vs (after the change).