@liancheng 2016-11-01
@tejasapatil For `collect_set` and `collect_list`, we'll simply migrate them to `TypedImperativeAggregate` and so that they become Spark native aggregate functions. We can also handle other built-in Hive UDAFs that share the same issue in this approach.

In the long run, this PR is mostly useful for running legacy UDAFs developed by the users rather than Hive built-in ones since we'd prefer to migrate all the built-in ones to Spark at last.