@liancheng 2016-11-01
I found that I'm handling bridged UDAFs properly, which caused a few test failures. Working on it.