@liancheng 2016-11-01
@tejasapatil Another point that I'd like to add is that even if the performance for a single UDAF like `GenericUDAFCollectList` regresses, you still have performance gains if such UDAFs are used together with other aggregate functions within a single query. Before this PR, all the aggregate functions have to be evaluated without partial aggregation support in this case.