@liancheng 2016-11-01
It turned out that I didn't initialize Hive UDAF evaluators properly.

Quoted from commit message of my previous commit:

> Hive UDAFs are sensitive to aggregation mode, and must be initialized with proper modes before being used. Basically, it means that you can't use an evaluator initialized with mode `PARTIAL1` to merge two aggregation states (although it still works for aggregate functions whose partial result type is the same as the final result type).