@liancheng 2016-11-02
OK, now it's ready for review and merge.

cc @yhuai @JoshRosen @cloud-fan