@liancheng 2016-11-10
The last build failure was because of a logical conflict between this PR and the master branch. Resolving it.