@yhuai 2016-11-16
Code changes looks good to me. Let's also do a benchmark to sanity check our implementation.