@yhuai 2016-11-16
LGTM. Merging to master.